+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.3.0 - 2024-03-27
 
 -   Fix incorrect icon on the `renpyWarp.warpToFile` command.
 -   Allow `renpyWarp.launch` to be called for a workspace folder context
--   Add a run button for the `renpyWarp.launch` command
+-   Add a status bar button for the `renpyWarp.launch` command
 
 ## 0.2.0 - 2024-03-27